callgraph: Add a CallGraph class and use it in views
The ICallGraphProvider, instead of managing the elements and CCTs will
now create and populate a CallGraph object that can be used in views.

This will allow various views to use instances of callgraph from a
given provider, for different time ranges for instance.

This adds a new flamegraph view that responds to selection instead of
showing everything. Both views can be shown together.

Change-Id: I2165a53a00a2b6e555cc62dc6d9e6550c7e39d01
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>