vm: add a VM overhead callstack
This analysis adds a callstack under the guest's running thread
to show when the thread is not actively running (either in VMM
mode or preempted on the host), as well as the reason for which
it was not running.

The resulting flamegraph allows to see the reason of the overhead.

Change-Id: Ic6ef110cdb5e6067650c183a34433322720503a5