Scripting: Add scripting capabilities to Trace Compass
This adds a few scripting plugins to the incubator, the scripting uses
the EASE project from Eclipse.

With this patch, the EASE feature works only from the Trace Compass
sources.

The modules added are:

* Analysis: Create an analysis with a state system from a script
* DataProvider: Create data providers from analyses, either using data
to define what to draw, or using scrited functions to directly get the
data.
* Filters: A script can generate a filter regex string to filter the
views accordingly
* Utils: Some java classes like Lists are not transparently passed
between script and Java, so this module adds wrappers around those
classes.
* View: Allow to open views with scripted data providers.

Change-Id: I1b402a2dd355d253f74ebf650ad865ea05a986a2
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>